Add search bar on results page. Improve navigation via breadcrumbs
