+++ client/app/queue/QueueApp.jsx
@@ -24,6 +24,7 @@ import SubmitDecisionView from './SubmitDecisionView';
 import SelectDispositionsView from './SelectDispositionsView';
 import AddEditIssueView from './AddEditIssueView';
 import SelectRemandReasonsView from './SelectRemandReasonsView';
+import SearchBar from './SearchBar';
 
 import { LOGO_COLORS } from '../constants/AppConstants';
 import { DECISION_TYPES } from './constants';
@@ -32,6 +33,11 @@ const appStyling = css({ paddingTop: '3rem' });
 
 class QueueApp extends React.PureComponent {
   routedSearchHome = () => <QueueLoadingScreen {...this.props}>
+    { this.props.searchedAppeals.length > 0 &&
+      <SearchBar
+        feedbackUrl={this.props.feedbackUrl}
+        shouldUseQueueCaseSearch={this.props.featureToggles.queue_case_search} />
+    }
     <CaseListView
       backLinkTarget="/"
       backLinkText="< Back to Case Search"
@@ -180,7 +186,8 @@ QueueApp.propTypes = {
 const mapStateToProps = (state) => ({
   ..._.pick(state.caseSelect, 'caseSelectCriteria.searchQuery'),
   ..._.pick(state.queue.loadedQueue, 'appeals'),
-  reviewActionType: state.queue.stagedChanges.taskDecision.type
+  reviewActionType: state.queue.stagedChanges.taskDecision.type,
+  searchedAppeals: state.caseList.receivedAppeals
 });
 
 export default connect(mapStateToProps)(QueueApp);
