+++ client/app/queue/SearchEnabledView.jsx
@@ -1,52 +1,11 @@
-import { css } from 'glamor';
 import PropTypes from 'prop-types';
 import React from 'react';
 import { connect } from 'react-redux';
 
-import CaseListSearch from './CaseListSearch';
 import CaseListView from './CaseListView';
-import CaseSelectSearch from '../reader/CaseSelectSearch';
-
-const searchStyling = (isRequestingAppealsUsingVeteranId) => css({
-  '.section-search': {
-    '& .usa-alert-info, & .usa-alert-error': {
-      marginBottom: '1.5rem',
-      marginTop: 0
-    },
-    '& .cf-search-input-with-close': {
-      marginLeft: `calc(100% - ${isRequestingAppealsUsingVeteranId ? '60' : '56.5'}rem)`
-    },
-    '& .cf-submit': {
-      width: '10.5rem'
-    }
-  }
-});
+import SearchBar from './SearchBar';
 
 class SearchEnabledView extends React.PureComponent {
-  searchBar() {
-    if (this.props.shouldUseQueueCaseSearch) {
-      // Do not draw the search bar in the top left when the search caused an error.
-      if (this.props.errorType) {
-        return;
-      }
-
-      return <div className="section-search" {...searchStyling(this.props.isRequestingAppealsUsingVeteranId)}>
-        <CaseListSearch />
-      </div>;
-    }
-
-    return <CaseSelectSearch
-      navigateToPath={(path) => {
-        const redirectUrl = encodeURIComponent(window.location.pathname);
-
-        location.href = `/reader/appeal${path}?queue_redirect_url=${redirectUrl}`;
-      }}
-      alwaysShowCaseSelectionModal
-      feedbackUrl={this.props.feedbackUrl}
-      searchSize="big"
-      styling={searchStyling(this.props.isRequestingAppealsUsingVeteranId)} />;
-  }
-
   render() {
     const {
       appeals,
@@ -55,7 +14,7 @@ class SearchEnabledView extends React.PureComponent {
     } = this.props;
 
     return <React.Fragment>
-      { this.searchBar() }
+      <SearchBar feedbackUrl={this.props.feedbackUrl} shouldUseQueueCaseSearch={this.props.shouldUseQueueCaseSearch} />
       { shouldUseQueueCaseSearch && (appeals.length > 0 || errorType) ? <CaseListView /> : this.props.children }
     </React.Fragment>;
   }
@@ -68,8 +27,7 @@ SearchEnabledView.propTypes = {
 
 const mapStateToProps = (state) => ({
   appeals: state.caseList.receivedAppeals,
-  errorType: state.caseList.search.errorType,
-  isRequestingAppealsUsingVeteranId: state.caseList.isRequestingAppealsUsingVeteranId
+  errorType: state.caseList.search.errorType
 });
 
 export default connect(mapStateToProps)(SearchEnabledView);
